Convertion to title case with lower case for some classes of words.

In written Portuguese, when converting to title case, it is not usual to keep in title case some words, like prepositions, conjunctions, articles and some kinds of pronouns. This functions locates those cases and converts them to lower case. Based on gist ToTitleCasePT https://gist.github.com/jjesusfilho/454192db8356eb9c486a02698338221a by https://github.com/jjesusfilho

Usage

ToTitleCasePT(string, ...)

Arguments

string

vector of characters to be converted to title case but with connector words (one-word prepositions and conjunctions) and articles (both definite and indefinite) and some pronouns in lower case.

Value

vector of characters with the same dimension of string

Details

The current list of words converted to lower case is:

articles

o(s), a(s), um, uma(s), uns

pronouns

me, mim, meu(s), minha(s)

te, ti, teu(s), tua(s)

lhe(s), seu(s), sua(s)

nos, nosso(a)(s)

vos, vosso(a)(s)

prepositions, their contractions and combinations

prepositions

a, ante, até, após, com, contra, de, desde, em, entre, para, perante, por, sem, sob, sobre, trás

contractions

à(s), do(a)(s), no(a)(s), pelo(a)(s), pro(a)(s) (informal language)

combinations

ao(s)

conjunctions

conforme, conquanto, contudo, durante, embora, enquanto, então, entretanto, exceto, logo, mas, nem, ou, ora, pois, porém, porque, porquanto, portanto, quando, quanto, que, se, senão, todavia

The above list is far from complete or exaustive, mainly due to the absence of the accidental prepositions and conjuntions, which are words that are not originally prepositions or conjunctions, but can play those roles in some contexts, like, for instance segundo, which can mean either the numeral second or the prepositional expression acording to.