Alocação Otimizada de Unidades de Coleta (UCs) a Agências

Esta função realiza a alocação otimizada de Unidades de Coleta (UCs) a agências, com o objetivo de minimizar os custos totais de deslocamento e operação. A alocação leva em consideração restrições de capacidade das agências, custos de deslocamento (combustível, tempo de viagem e diárias), custos fixos das agências e custos de treinamento.

Usage

alocar_ucs(
  ucs,
  agencias = NULL,
  custo_litro_combustivel = 6,
  custo_hora_viagem = 10,
  kml = 10,
  valor_diaria = 335,
  custo_fixo = 500,
  dias_treinamento = 5.5,
  dist_diaria_km = 100,
  agencias_treinadas = NULL,
  agencias_treinamento = NULL,
  distancias_ucs,
  distancias_agencias,
  min_uc_agencia = 1,
  max_uc_agencia = NULL,
  semi_centralizada = NULL,
  adicional_troca_jurisdicao = 0,
  resultado_completo = FALSE
)

Arguments

ucs

Um tibble ou data.frame contendo informações sobre as UCs, incluindo:

uc: Código único da UC.

agencia_codigo: Código da agência à qual a UC está atualmente alocada.

dias_coleta: Número de dias de coleta na UC.

viagens: Número de viagens necessárias para a coleta na UC.

agencias

Um tibble ou data.frame contendo informações sobre as agências selecionáveis, incluindo:

agencia_codigo: Código único da agência.

custo_litro_combustivel

Custo do combustível por litro (em R$). Padrão: 6.

custo_hora_viagem

Custo de cada hora de viagem (em R$). Padrão: 10.

kml

Consumo médio de combustível do veículo (em km/l). Padrão: 10.

valor_diaria

Valor da diária para deslocamentos (em R$). Padrão: 335.

custo_fixo

Custo fixo mensal da agência, como salário do entrevistador, custo de monitoramento, etc. (em R$). Padrão: 500.

dias_treinamento

Número de dias/diárias para treinamento. Padrão: 5.5

dist_diaria_km

Distância mínima (em km) para que seja paga uma diária, no caso de UC na jurisdição da agência. No caso da UC fora da jurisdição da agência, quando a distância for menor do que dist_diaria_km, se paga meia diária, calculada como 1/2 do valor das diárias inteiras. Padrão 100

agencias_treinadas

(Opcional) Um vetor de caracteres com os códigos das agências que já foram treinadas e não terão custo de treinamento. O custo dos APMs contratados nessas agências ainda será incluído no plano de otimização. Padrão: NULL.

agencias_treinamento

Código da(s) agência(s) onde o treinamento será realizado.

distancias_ucs

Um tibble ou data.frame com as distâncias entre UCs e agências, incluindo:

uc: Código da UC.

agencia_codigo: Código da agência.

distancia_km: Distância em quilômetros entre a UC e a agência.

duracao_horas: Duração da viagem em horas entre a UC e a agência

diaria_municipio: Indica se é necessária uma diária para deslocamento entre a UC e a agência, considerando o município da UC

distancias_agencias

Um tibble ou data.frame com as distâncias entre as agências, incluindo:

agencia_codigo_orig: Código da agência de origem

agencia_codigo_dest: Código da agência de destino

distancia_km: Distância em quilômetros entre a agência de origem e a de destino

duracao_horas: Duração da viagem em horas entre a agência de origem e a de destino

min_uc_agencia

Número mínimo de UCs por agência. Só válido para agências não treinadas. Padrão: 1.

max_uc_agencia

Número máximo de UCs por agência. Padrão: NULL (ilimitado).

semi_centralizada

(Opcional) Um vetor de caracteres com os códigos das agências que não terão limite máximo de UCs alocadas. Padrão: NULL.

adicional_troca_jurisdicao

Custo adicional quando há troca de agência de coleta. Padrão 0

resultado_completo

(Opcional) Um valor lógico indicando se deve ser retornado um resultado mais completo, incluindo informações sobre todas as combinações de UCs e agências. Padrão: FALSE.

Value

Uma lista contendo:

resultado_ucs_otimo: Um tibble com as UCs e suas alocações otimizadas, incluindo custos de deslocamento.

resultado_ucs_jurisdicao: Um tibble com as UCs e suas alocações originais (jurisdição), incluindo custos de deslocamento.

resultado_agencias_otimo: Um tibble com as agências e suas alocações otimizadas, incluindo custos fixos, custos de deslocamento e número de UCs alocadas.

resultado_agencias_jurisdicao: Um tibble com as agências e suas alocações originais (jurisdição), incluindo custos fixos, custos de deslocamento e número de UCs alocadas.

ucs_agencias_todas (opcional): Um tibble com todas as combinações de UCs e agências, incluindo distâncias, custos e informações sobre diárias (retornado apenas se resultado_completo for TRUE).