Calcula Distâncias Entre Origens e Destinos

Esta função calcula a distância e duração entre conjuntos de pontos de origem e destino usando o serviço OSRM.

Usage

calcula_distancias(destinos, origens, nmax = 2000)

Arguments

destinos

Um objeto sf representando os pontos de destino.

origens

Um objeto sf representando os pontos de origem.

nmax

Número máximo de pares origem-destino para cada requisição ao serviço OSRM. Utilizado para evitar o limite de número de pares por requisição. Padrão: 2000.

Value

Um data.frame com as distâncias (em quilômetros) e durações (em horas) entre cada par de origem e destino, além dos atributos originais de origens e destinos.

Details

A função utiliza o pacote osrm para calcular as distâncias e durações entre os pontos de origem e destino. O parâmetro nmax permite dividir as requisições ao serviço OSRM em lotes menores, evitando o erro "Too many table coordinates" que ocorre quando o número de pares origem-destino é muito grande.